Remove GameViewModel.kt and rename GameViewModelRefactor to GameViewModel
