+++ app/src/main/java/pl/kazoroo/tavernFarkle/MainActivity.kt
@@ -25,7 +25,7 @@ import pl.kazoroo.tavernFarkle.core.presentation.CoinsViewModel
 import pl.kazoroo.tavernFarkle.core.presentation.navigation.Navigation
 import pl.kazoroo.tavernFarkle.di.DependencyContainer
 import pl.kazoroo.tavernFarkle.di.TavernFarkleApp
-import pl.kazoroo.tavernFarkle.game.presentation.game.GameViewModelRefactor
+import pl.kazoroo.tavernFarkle.game.presentation.game.GameViewModel
 import pl.kazoroo.tavernFarkle.game.presentation.mainmenu.MainMenuViewModel
 import pl.kazoroo.tavernFarkle.game.presentation.sound.SoundPlayer
 import pl.kazoroo.tavernFarkle.game.presentation.splashscreen.StartingScreenViewModel
@@ -98,7 +98,7 @@ class MainActivity : ComponentActivity() {
             val inventoryViewModel = viewModel<InventoryViewModel>(
                 factory = dependencyContainer.inventoryViewModelFactory
             )
-            val gameViewModel = viewModel<GameViewModelRefactor>(
+            val gameViewModel = viewModel<GameViewModel>(
                 factory = dependencyContainer.gameViewModelFactory
             )
