+++ app/src/main/java/pl/kazoroo/tavernFarkle/core/presentation/navigation/Navigation.kt
@@ -8,7 +8,7 @@ import androidx.navigation.compose.rememberNavController
 import pl.kazoroo.tavernFarkle.core.presentation.CoinsViewModel
 import pl.kazoroo.tavernFarkle.di.DependencyContainer
 import pl.kazoroo.tavernFarkle.game.presentation.game.GameScreen
-import pl.kazoroo.tavernFarkle.game.presentation.game.GameViewModelRefactor
+import pl.kazoroo.tavernFarkle.game.presentation.game.GameViewModel
 import pl.kazoroo.tavernFarkle.game.presentation.mainmenu.MainMenuScreen
 import pl.kazoroo.tavernFarkle.game.presentation.mainmenu.MainMenuViewModel
 import pl.kazoroo.tavernFarkle.multiplayer.presentation.LobbyScreen
@@ -26,7 +26,7 @@ fun Navigation(
     coinsViewModel: CoinsViewModel,
     inventoryViewModel: InventoryViewModel,
     settingsViewModel: SettingsViewModel,
-    gameViewModel: GameViewModelRefactor
+    gameViewModel: GameViewModel
 ) {
     val navController = rememberNavController()
